Fixed pygame warning when switching discs

+++ main.py
@@ -38,7 +38,7 @@ def DrawBag(discSlot, hoverCheck):
         pygame.draw.circle(screen,
                            color.LIGHT_GREY,
                            [int(discSlot.x + 0.5 * discSlot.r), int(discSlot.y + 0.5 * discSlot.r)],
-                           (discSlot.r + 10) * 0.5,
+                           int((discSlot.r + 10) * 0.5),
                            width=0)
     DrawCircle(discSlot, discSlot.color)
     discTypeText = pygame.font.Font('freesansbold.ttf', 12)
